# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: College Management System API for Tamil Nadu
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /library-books:
    get:
      operationId: listLibraryBooks
      tags: [library]
      summary: List library books
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Book list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryBook"
    post:
      operationId: createLibraryBook
      tags: [library]
      summary: Add book to library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLibraryBookBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryBook"
  /library-books/{id}:
    get:
      operationId: getLibraryBook
      tags: [library]
      summary: Get book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryBook"
    patch:
      operationId: updateLibraryBook
      tags: [library]
      summary: Update book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLibraryBookBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryBook"
    delete:
      operationId: deleteLibraryBook
      tags: [library]
      summary: Delete book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /library-issued:
    get:
      operationId: listLibraryIssued
      tags: [library]
      summary: List issued books
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Issued list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryIssuedBook"
    post:
      operationId: issueLibraryBook
      tags: [library]
      summary: Issue book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueLibraryBookBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryIssuedBook"
  /library-issued/{id}:
    patch:
      operationId: updateLibraryIssued
      tags: [library]
      summary: Update issued book (return/renew)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLibraryIssuedBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryIssuedBook"

  # ── Events ──
...
  /assets:
    get:
      operationId: listAssets
      tags: [inventory]
      summary: List assets
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Asset list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [inventory]
      summary: Create asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [inventory]
      summary: Get asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    patch:
      operationId: updateAsset
      tags: [inventory]
      summary: Update asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssetBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [inventory]
      summary: Delete asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /store-items:
    get:
      operationId: listStoreItems
      tags: [inventory]
      summary: List store items
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Item list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoreItem"
    post:
      operationId: createStoreItem
      tags: [inventory]
      summary: Create store item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreItem"
  /store-items/{id}:
    patch:
      operationId: updateStoreItem
      tags: [inventory]
      summary: Update store item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoreItemBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreItem"
    delete:
      operationId: deleteStoreItem
      tags: [inventory]
      summary: Delete store item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # ── Timetable ──
  /timetable:
    get:
      operationId: listTimetable
      tags: [academics]
      summary: List timetable entries
      parameters:
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: semester
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Timetable
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimetableEntry"
    post:
      operationId: createTimetableEntry
      tags: [academics]
      summary: Create timetable entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimetableEntryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimetableEntry"
  /timetable/{id}:
    patch:
      operationId: updateTimetableEntry
      tags: [academics]
      summary: Update timetable entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTimetableEntryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimetableEntry"
    delete:
      operationId: deleteTimetableEntry
      tags: [academics]
      summary: Delete timetable entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # ── Assignments ──
...
  /staff-leaves:
    get:
      operationId: listStaffLeaves
      tags: [staff]
      summary: List staff leaves
      parameters:
        - name: staffId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leave list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffLeave"
    post:
      operationId: createStaffLeave
      tags: [staff]
      summary: Apply for leave
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffLeaveBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffLeave"
  /staff-leaves/{id}:
    get:
      operationId: getStaffLeave
      tags: [staff]
      summary: Get leave
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leave
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffLeave"
    patch:
      operationId: updateStaffLeave
      tags: [staff]
      summary: Update leave (approve/reject)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffLeaveBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffLeave"
components:
  schemas:
...
    LibraryBook:
      type: object
      properties:
        id:
          type: integer
        isbn:
          type: string
          nullable: true
        title:
          type: string
        author:
          type: string
        publisher:
          type: string
          nullable: true
        edition:
          type: string
          nullable: true
        category:
          type: string
        subject:
          type: string
          nullable: true
        shelfLocation:
          type: string
          nullable: true
        totalCopies:
          type: integer
        availableCopies:
          type: integer
        price:
          type: number
          nullable: true
        yearOfPublication:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, author, category, totalCopies, availableCopies, status, createdAt]
    CreateLibraryBookBody:
      type: object
      properties:
        isbn:
          type: string
        title:
          type: string
        author:
          type: string
        publisher:
          type: string
        edition:
          type: string
        category:
          type: string
        subject:
          type: string
        shelfLocation:
          type: string
        totalCopies:
          type: integer
        availableCopies:
          type: integer
        price:
          type: number
        yearOfPublication:
          type: string
        status:
          type: string
      required: [title, author, category]
    UpdateLibraryBookBody:
      type: object
      properties:
        isbn:
          type: string
        title:
          type: string
        author:
          type: string
        publisher:
          type: string
        edition:
          type: string
        category:
          type: string
        subject:
          type: string
        shelfLocation:
          type: string
        totalCopies:
          type: integer
        availableCopies:
          type: integer
        price:
          type: number
        yearOfPublication:
          type: string
        status:
          type: string
    LibraryIssuedBook:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        memberId:
          type: integer
        memberType:
          type: string
        issueDate:
          type: string
        dueDate:
          type: string
        returnDate:
          type: string
          nullable: true
        fineAmount:
          type: number
          nullable: true
        fineStatus:
          type: string
          nullable: true
        status:
          type: string
        remarks:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, bookId, memberId, memberType, issueDate, dueDate, status, createdAt]
    IssueLibraryBookBody:
      type: object
      properties:
        bookId:
          type: integer
        memberId:
          type: integer
        memberType:
          type: string
        issueDate:
          type: string
        dueDate:
          type: string
        remarks:
          type: string
      required: [bookId, memberId, memberType, issueDate, dueDate]
    UpdateLibraryIssuedBody:
      type: object
      properties:
        returnDate:
          type: string
        fineAmount:
          type: number
        fineStatus:
          type: string
        status:
          type: string
        remarks:
          type: string

    # ── Event Schemas ──
...
    Asset:
      type: object
      properties:
        id:
          type: integer
        assetTag:
          type: string
        name:
          type: string
        category:
          type: string
        departmentId:
          type: integer
          nullable: true
        location:
          type: string
          nullable: true
        purchaseDate:
          type: string
          nullable: true
        purchasePrice:
          type: number
          nullable: true
        vendor:
          type: string
          nullable: true
        warrantyExpiry:
          type: string
          nullable: true
        condition:
          type: string
        assignedTo:
          type: string
          nullable: true
        serialNumber:
          type: string
          nullable: true
        status:
          type: string
        remarks:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, assetTag, name, category, condition, status, createdAt]
    CreateAssetBody:
      type: object
      properties:
        assetTag:
          type: string
        name:
          type: string
        category:
          type: string
        departmentId:
          type: integer
        location:
          type: string
        purchaseDate:
          type: string
        purchasePrice:
          type: number
        vendor:
          type: string
        warrantyExpiry:
          type: string
        condition:
          type: string
        assignedTo:
          type: string
        serialNumber:
          type: string
        status:
          type: string
        remarks:
          type: string
      required: [assetTag, name, category]
    UpdateAssetBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        departmentId:
          type: integer
        location:
          type: string
        purchaseDate:
          type: string
        purchasePrice:
          type: number
        vendor:
          type: string
        warrantyExpiry:
          type: string
        condition:
          type: string
        assignedTo:
          type: string
        serialNumber:
          type: string
        status:
          type: string
        remarks:
          type: string
    StoreItem:
      type: object
      properties:
        id:
          type: integer
        itemCode:
          type: string
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        currentStock:
          type: integer
        minimumStock:
          type: integer
        unitPrice:
          type: number
          nullable: true
        lastRestockDate:
          type: string
          nullable: true
        supplier:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, itemCode, name, category, unit, currentStock, minimumStock, status, createdAt]
    CreateStoreItemBody:
      type: object
      properties:
        itemCode:
          type: string
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        currentStock:
          type: integer
        minimumStock:
          type: integer
        unitPrice:
          type: number
        supplier:
          type: string
        status:
          type: string
      required: [itemCode, name, category]
    UpdateStoreItemBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        currentStock:
          type: integer
        minimumStock:
          type: integer
        unitPrice:
          type: number
        lastRestockDate:
          type: string
        supplier:
          type: string
        status:
          type: string

    # ── Academic Schemas ──
    TimetableEntry:
      type: object
      properties:
        id:
          type: integer
        departmentId:
          type: integer
        semester:
          type: integer
        dayOfWeek:
          type: string
        periodNumber:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        subjectId:
          type: integer
          nullable: true
        staffId:
          type: integer
          nullable: true
        room:
          type: string
          nullable: true
        section:
          type: string
          nullable: true
        academicYear:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, departmentId, semester, dayOfWeek, periodNumber, startTime, endTime, academicYear, createdAt]
    CreateTimetableEntryBody:
      type: object
      properties:
        departmentId:
          type: integer
        semester:
          type: integer
        dayOfWeek:
          type: string
        periodNumber:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        subjectId:
          type: integer
        staffId:
          type: integer
        room:
          type: string
        section:
          type: string
        academicYear:
          type: string
      required: [departmentId, semester, dayOfWeek, periodNumber, startTime, endTime, academicYear]
    UpdateTimetableEntryBody:
      type: object
      properties:
        dayOfWeek:
          type: string
        periodNumber:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        subjectId:
          type: integer
        staffId:
          type: integer
        room:
          type: string
        section:
          type: string
...
    StaffLeave:
      type: object
      properties:
        id:
          type: integer
        staffId:
          type: integer
        leaveType:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        totalDays:
          type: integer
        reason:
          type: string
        status:
          type: string
        approvedBy:
          type: string
          nullable: true
        remarks:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, staffId, leaveType, startDate, endDate, totalDays, reason, status, createdAt]
    CreateStaffLeaveBody:
      type: object
      properties:
        staffId:
          type: integer
        leaveType:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        totalDays:
          type: integer
        reason:
          type: string
        remarks:
          type: string
      required: [staffId, leaveType, startDate, endDate, totalDays, reason]
    UpdateStaffLeaveBody:
      type: object
      properties:
        status:
          type: string
        approvedBy:
          type: string
        remarks:
          type: string