# lib/api-zod/src/generated/types/libraryIssuedBook.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * College Management System API for Tamil Nadu
 * OpenAPI spec version: 0.1.0
 */

export interface LibraryIssuedBook {
  id: number;
  bookId: number;
  memberId: number;
  memberType: string;
  issueDate: string;
  dueDate: string;
  returnDate?: string | null;
  fineAmount?: number | null;
  fineStatus?: string | null;
  status: string;
  remarks?: string | null;
  createdAt: Date;
}